refactor & add some comments
